+++ tests/Bridge/Elasticsearch/DataProvider/CollectionDataProviderTest.php
@@ -17,12 +17,14 @@ use ApiPlatform\Core\Bridge\Elasticsearch\Api\IdentifierExtractorInterface;
 use ApiPlatform\Core\Bridge\Elasticsearch\DataProvider\CollectionDataProvider;
 use ApiPlatform\Core\Bridge\Elasticsearch\DataProvider\Extension\RequestBodySearchCollectionExtensionInterface;
 use ApiPlatform\Core\Bridge\Elasticsearch\DataProvider\Paginator;
+use ApiPlatform\Core\Bridge\Elasticsearch\DataProvider\PaginatorFactory;
 use ApiPlatform\Core\Bridge\Elasticsearch\Exception\IndexNotFoundException;
 use ApiPlatform\Core\Bridge\Elasticsearch\Exception\NonUniqueIdentifierException;
 use ApiPlatform\Core\Bridge\Elasticsearch\Metadata\Document\DocumentMetadata;
 use ApiPlatform\Core\Bridge\Elasticsearch\Metadata\Document\Factory\DocumentMetadataFactoryInterface;
 use ApiPlatform\Core\DataProvider\CollectionDataProviderInterface;
 use ApiPlatform\Core\DataProvider\Pagination;
+use ApiPlatform\Core\Exception\InvalidArgumentException;
 use ApiPlatform\Core\Exception\ResourceClassNotFoundException;
 use ApiPlatform\Core\Metadata\Resource\Factory\ResourceMetadataFactoryInterface;
 use ApiPlatform\Core\Metadata\Resource\ResourceMetadata;
@@ -35,10 +37,12 @@ use ApiPlatform\Core\Tests\ProphecyTrait;
 use Elasticsearch\Client;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
+use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 
 class CollectionDataProviderTest extends TestCase
 {
+    use ExpectDeprecationTrait;
     use ProphecyTrait;
 
     public function testConstruct()
@@ -49,10 +53,28 @@ class CollectionDataProviderTest extends TestCase
                 $this->prophesize(Client::class)->reveal(),
                 $this->prophesize(DocumentMetadataFactoryInterface::class)->reveal(),
                 $this->prophesize(IdentifierExtractorInterface::class)->reveal(),
-                $this->prophesize(DenormalizerInterface::class)->reveal(),
+                new PaginatorFactory($this->prophesize(DenormalizerInterface::class)->reveal()),
                 new Pagination($this->prophesize(ResourceMetadataFactoryInterface::class)->reveal()),
                 $this->prophesize(ResourceMetadataFactoryInterface::class)->reveal()
             )
+        );
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testConstructWithoutDenormalizerInterfaceOrPaginatorFactoryInterfaceInstanceIsForbidden()
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The fourth argument of the ApiPlatform\\Core\\Bridge\\Elasticsearch\\DataProvider\\CollectionDataProvider::__construct method must be a ApiPlatform\\Core\\DataProvider\\PaginatorFactoryInterface instance.');
+
+        new CollectionDataProvider(
+            $this->prophesize(Client::class)->reveal(),
+            $this->prophesize(DocumentMetadataFactoryInterface::class)->reveal(),
+            $this->prophesize(IdentifierExtractorInterface::class)->reveal(),
+            new \stdClass(), // @phpstan-ignore-line
+            new Pagination($this->prophesize(ResourceMetadataFactoryInterface::class)->reveal()),
+            $this->prophesize(ResourceMetadataFactoryInterface::class)->reveal()
         );
     }
 
@@ -78,7 +100,7 @@ class CollectionDataProviderTest extends TestCase
             $this->prophesize(Client::class)->reveal(),
             $documentMetadataFactoryProphecy->reveal(),
             $identifierExtractorProphecy->reveal(),
-            $this->prophesize(DenormalizerInterface::class)->reveal(),
+            new PaginatorFactory($this->prophesize(DenormalizerInterface::class)->reveal()),
             new Pagination($this->prophesize(ResourceMetadataFactoryInterface::class)->reveal()),
             $resourceMetadataFactoryProphecy->reveal()
         );
@@ -94,6 +116,7 @@ class CollectionDataProviderTest extends TestCase
     {
         $context = [
             'groups' => ['custom'],
+            'resourceClass' => Foo::class,
         ];
 
         $documentMetadataFactoryProphecy = $this->prophesize(DocumentMetadataFactoryInterface::class);
@@ -164,6 +187,103 @@ class CollectionDataProviderTest extends TestCase
 
         $requestBodySearchCollectionExtensionProphecy = $this->prophesize(RequestBodySearchCollectionExtensionInterface::class);
         $requestBodySearchCollectionExtensionProphecy->applyToCollection([], Foo::class, 'get', $context)->willReturn([])->shouldBeCalled();
+
+        $collectionDataProvider = new CollectionDataProvider(
+            $clientProphecy->reveal(),
+            $documentMetadataFactoryProphecy->reveal(),
+            $this->prophesize(IdentifierExtractorInterface::class)->reveal(),
+            new PaginatorFactory($denormalizer = $this->prophesize(DenormalizerInterface::class)->reveal()),
+            new Pagination($resourceMetadataFactoryProphecy->reveal(), ['items_per_page' => 2]),
+            $resourceMetadataFactoryProphecy->reveal(),
+            [$requestBodySearchCollectionExtensionProphecy->reveal()]
+        );
+
+        self::assertEquals(
+            new Paginator($denormalizer, $documents, Foo::class, 2, 0, $context),
+            $collectionDataProvider->getCollection(Foo::class, 'get', $context)
+        );
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testGetCollectionWithDeprecatedDenormalizerDependency()
+    {
+        $context = [
+            'groups' => ['custom'],
+            'resourceClass' => Foo::class,
+        ];
+
+        $documentMetadataFactoryProphecy = $this->prophesize(DocumentMetadataFactoryInterface::class);
+        $documentMetadataFactoryProphecy->create(Foo::class)->willReturn(new DocumentMetadata('foo'))->shouldBeCalled();
+
+        $resourceMetadataFactoryProphecy = $this->prophesize(ResourceMetadataFactoryInterface::class);
+        $resourceMetadataFactoryProphecy->create(Foo::class)->willReturn(new ResourceMetadata());
+
+        $documents = [
+            'took' => 15,
+            'time_out' => false,
+            '_shards' => [
+                'total' => 5,
+                'successful' => 5,
+                'skipped' => 0,
+                'failed' => 0,
+            ],
+            'hits' => [
+                'total' => 4,
+                'max_score' => 1,
+                'hits' => [
+                    [
+                        '_index' => 'foo',
+                        '_type' => '_doc',
+                        '_id' => '1',
+                        '_score' => 1,
+                        '_source' => [
+                            'id' => 1,
+                            'name' => 'Kilian',
+                            'bar' => 'Jornet',
+                        ],
+                    ],
+                    [
+                        '_index' => 'foo',
+                        '_type' => '_doc',
+                        '_id' => '2',
+                        '_score' => 1,
+                        '_source' => [
+                            'id' => 2,
+                            'name' => 'François',
+                            'bar' => 'D\'Haene',
+                        ],
+                    ],
+                ],
+            ],
+        ];
+
+        $clientProphecy = $this->prophesize(Client::class);
+        $clientProphecy
+            ->search(
+                Argument::allOf(
+                    Argument::withEntry('index', 'foo'),
+                    Argument::withEntry('type', DocumentMetadata::DEFAULT_TYPE),
+                    Argument::withEntry('body', Argument::allOf(
+                        Argument::withEntry('size', 2),
+                        Argument::withEntry('from', 0),
+                        Argument::withEntry('query', Argument::allOf(
+                            Argument::withEntry('match_all', Argument::type(\stdClass::class)),
+                            Argument::size(1)
+                        )),
+                        Argument::size(3)
+                    )),
+                    Argument::size(3)
+                )
+            )
+            ->willReturn($documents)
+            ->shouldBeCalled();
+
+        $requestBodySearchCollectionExtensionProphecy = $this->prophesize(RequestBodySearchCollectionExtensionInterface::class);
+        $requestBodySearchCollectionExtensionProphecy->applyToCollection([], Foo::class, 'get', $context)->willReturn([])->shouldBeCalled();
+
+        $this->expectDeprecation('Injecting a Symfony\Component\Serializer\Normalizer\DenormalizerInterface instance as fourth argument has been deprecated since API Platform 2.6 and will be removed in API Platform 3. Inject a ApiPlatform\Core\DataProvider\PaginatorFactoryInterface instance instead.');
 
         $collectionDataProvider = new CollectionDataProvider(
             $clientProphecy->reveal(),